comm: Rename a lot of state pointer variables
We are going to split `struct udp' into a bunch of substructures with
their own types, to assist with reuse of the udp code in a new comm.
This is going to involve members at different levels of abstraction
being accessed through different pointer variable names.

So change the names now to aliases of the standard `struct udp *st'.
This will much reduce noise in subsequent patches.

Also rename udp.c's MAX_SOCKETS to UDP_MAX_SOCKETS.

No functional change in this patch.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>